services/vanna/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
vanna==0.5.5
psycopg[binary]==3.1.19
python-dotenv==1.0.1
groq==0.5.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
